+++ src/stitching/mergeSchemas.ts
@@ -34,6 +34,7 @@ import {
   isNamedType,
   parse,
   visit,
+  subscribe,
 } from 'graphql';
 import TypeRegistry from './TypeRegistry';
 import { IResolvers } from '../Interfaces';
@@ -51,7 +52,7 @@ import { checkResultAndHandleErrors } from './errors';
 
 export type MergeInfo = {
   delegate: (
-    type: 'query' | 'mutation',
+    type: 'query' | 'mutation' | 'subscription',
     fieldName: string,
     args: { [key: string]: any },
     context: { [key: string]: any },
@@ -83,6 +84,7 @@ export default function mergeSchemas({
   }
   let queryFields: GraphQLFieldMap<any, any> = {};
   let mutationFields: GraphQLFieldMap<any, any> = {};
+  let subscriptionFields: GraphQLFieldMap<any, any> = {};
 
   const typeRegistry = new TypeRegistry();
 
@@ -115,6 +117,7 @@ export default function mergeSchemas({
     typeRegistry.addSchema(schema);
     const queryType = schema.getQueryType();
     const mutationType = schema.getMutationType();
+    const subscriptionType = schema.getSubscriptionType();
 
     const typeMap = schema.getTypeMap();
     Object.keys(typeMap).forEach(typeName => {
@@ -123,7 +126,8 @@ export default function mergeSchemas({
         isNamedType(type) &&
         getNamedType(type).name.slice(0, 2) !== '__' &&
         type !== queryType &&
-        type !== mutationType
+        type !== mutationType &&
+        type !== subscriptionType
       ) {
         let newType;
         if (isCompositeType(type) || type instanceof GraphQLInputObjectType) {
@@ -171,6 +175,26 @@ export default function mergeSchemas({
         ...mutationType.getFields(),
       };
     }
+
+    if (subscriptionType) {
+      if (!fullResolvers.Subscription) {
+        fullResolvers.Subscription = {};
+      }
+      Object.keys(subscriptionType.getFields()).forEach(name => {
+        fullResolvers.Subscription[name] = {
+          subscribe: createDelegatingResolver(
+            mergeInfo,
+            'subscription',
+            name,
+          )
+        };
+      });
+
+      subscriptionFields = {
+        ...subscriptionFields,
+        ...subscriptionType.getFields(),
+      };
+    }
   });
 
   typeFragments.forEach(document => {
@@ -215,12 +239,22 @@ export default function mergeSchemas({
     });
   }
 
+  let subscription;
+  if (!isEmptyObject(subscriptionFields)) {
+    subscription = new GraphQLObjectType({
+      name: 'Subscription',
+      fields: () => fieldMapToFieldConfigMap(subscriptionFields, typeRegistry),
+    });
+  }
+
   typeRegistry.addType('Query', query);
   typeRegistry.addType('Mutation', mutation);
+  typeRegistry.addType('Subscription', subscription);
 
   let mergedSchema = new GraphQLSchema({
     query,
     mutation,
+    subscription,
     types: typeRegistry.getAllTypes(),
   });
 
@@ -236,7 +270,7 @@ export default function mergeSchemas({
 function createMergeInfo(typeRegistry: TypeRegistry): MergeInfo {
   return {
     delegate(
-      operation: 'query' | 'mutation',
+      operation: 'query' | 'mutation' | 'subscription',
       fieldName: string,
       args: { [key: string]: any },
       context: { [key: string]: any },
@@ -264,7 +298,7 @@ function createMergeInfo(typeRegistry: TypeRegistry): MergeInfo {
 
 function createDelegatingResolver(
   mergeInfo: MergeInfo,
-  operation: 'query' | 'mutation',
+  operation: 'query' | 'mutation' | 'subscription',
   fieldName: string,
 ): GraphQLFieldResolver<any, any> {
   return (root, args, context, info) => {
@@ -277,7 +311,7 @@ async function delegateToSchema(
   fragmentReplacements: {
     [typeName: string]: { [fieldName: string]: InlineFragmentNode };
   },
-  operation: 'query' | 'mutation',
+  operation: 'query' | 'mutation' | 'subscription',
   fieldName: string,
   args: { [key: string]: any },
   context: { [key: string]: any },
@@ -286,6 +320,8 @@ async function delegateToSchema(
   let type;
   if (operation === 'mutation') {
     type = schema.getMutationType();
+  } else if (operation === 'subscription') {
+    type = schema.getSubscriptionType();
   } else {
     type = schema.getQueryType();
   }
@@ -322,15 +358,26 @@ async function delegateToSchema(
       });
     }
 
-    const result = await execute(
-      schema,
-      graphqlDoc,
-      info.rootValue,
-      context,
-      variableValues,
-    );
+    if (operation === 'query' || operation === 'mutation') {
+      const result = await execute(
+        schema,
+        graphqlDoc,
+        info.rootValue,
+        context,
+        variableValues
+      );
+      return checkResultAndHandleErrors(result, info, fieldName);
+    }
 
-    return checkResultAndHandleErrors(result, info, fieldName);
+    if (operation === 'subscription') {
+      return subscribe(
+        schema,
+        graphqlDoc,
+        info.rootValue,
+        context,
+        variableValues
+      );
+    }
   }
 
   throw new Error('Could not forward to merged schema');
@@ -343,7 +390,7 @@ function createDocument(
   },
   type: GraphQLObjectType,
   rootFieldName: string,
-  operation: 'query' | 'mutation',
+  operation: 'query' | 'mutation' | 'subscription',
   selections: Array<FieldNode>,
   fragments: { [fragmentName: string]: FragmentDefinitionNode },
   variableDefinitions?: Array<VariableDefinitionNode>,