+++ src/test/testingSchemas.ts
@@ -13,6 +13,7 @@ import makeRemoteExecutableSchema, {
   Fetcher,
 } from '../stitching/makeRemoteExecutableSchema';
 import introspectSchema from '../stitching/introspectSchema';
+import { PubSub } from 'graphql-subscriptions';
 
 export type Property = {
   id: string;
@@ -244,6 +245,10 @@ const propertyRootTypeDefs = `
     bar: String
   }
 
+  input InputWithDefault {
+    test: String = "Foo"
+  }
+
   type Query {
     propertyById(id: ID!): Property
     properties(limit: Int): [Property!]
@@ -254,6 +259,7 @@ const propertyRootTypeDefs = `
     errorTest: String
     errorTestNonNull: String!
     relay: Query!
+    defaultInputTest(input: InputWithDefault!): String
   }
 `;
 
@@ -315,6 +321,10 @@ const propertyResolvers: IResolvers = {
 
     errorTestNonNull() {
       throw new Error('Sample error non-null!');
+    },
+
+    defaultInputTest(parent, { input }) {
+      return input.test;
     },
   },
   DateTime,
@@ -505,6 +515,35 @@ const bookingResolvers: IResolvers = {
   DateTime,
 };
 
+const subscriptionTypeDefs = `
+  type Notification{
+    text: String
+  }
+
+  type Query{
+    notifications: Notification
+  }
+
+  type Subscription{
+    notifications: Notification
+  }
+`;
+
+export const subscriptionPubSub = new PubSub();
+export const subscriptionPubSubTrigger = 'pubSubTrigger';
+
+const subscriptionResolvers: IResolvers = {
+  Query: {
+    notifications: (root: any) => ({ text: 'Hello world' }),
+  },
+  Subscription: {
+    notifications: {
+      subscribe: () =>
+        subscriptionPubSub.asyncIterator(subscriptionPubSubTrigger),
+    },
+  },
+};
+
 export const propertySchema: GraphQLSchema = makeExecutableSchema({
   typeDefs: propertyAddressTypeDefs,
   resolvers: propertyResolvers,
@@ -513,6 +552,11 @@ export const propertySchema: GraphQLSchema = makeExecutableSchema({
 export const bookingSchema: GraphQLSchema = makeExecutableSchema({
   typeDefs: bookingAddressTypeDefs,
   resolvers: bookingResolvers,
+});
+
+export const subscriptionSchema: GraphQLSchema = makeExecutableSchema({
+  typeDefs: subscriptionTypeDefs,
+  resolvers: subscriptionResolvers,
 });
 
 // Pretend this schema is remote